warga controller login

+++ database/migrations/2021_01_24_055657_create_wargas_table.php
@@ -17,7 +17,7 @@ class CreateWargasTable extends Migration
             $table->id();
             $table->string('nama');
             $table->string('alamat');	
-            $table->integer('nik')->unique();
+            $table->bigInteger('nik')->unique();
             $table->char('jenisKelamin',1 );	
             $table->date('tanggalLahir');
         });
